feat: add score UI
Remove UI from GameManage script

Refs: #35
