+++ Assets/UIManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 104e5dad7616d8446927b20d82fda1d0
