+++ Assets/_Scripts/GameManager.cs
@@ -14,17 +14,6 @@ namespace Assets._Scripts
                 Instance = this;
             else
                 Destroy(gameObject);
-
         }
-
-        public Slider soapTankSlider;
-        public TMP_Text soapTankText;
-
-        public void SetTankValue(int value)
-        {
-            soapTankSlider.value = value;
-            soapTankText.text = value.ToString();
-        }
-
     }
 }
