+++ Assets/_Scripts/HighFinder.cs
@@ -54,6 +54,10 @@ namespace Assets._Scripts
             localLinePos.z = 0;
             LocalHighLine.transform.position = localLinePos;
             LocalHighLine.transform.localScale = new Vector3(Math.Abs(bottomLeft.x) * 2 + 1, 0.05f, 1);
+
+            UIManager.Instance.SetBestScore((int)globalHighestY);
+            UIManager.Instance.SetCurrentScore((int)localHighestY);
+
         }
     }
 }
